refactor: refactor WorkerRequestHandler module
